Updated Query Params and added account/details endpoint

+++ src/users/user.service.ts
@@ -99,6 +99,9 @@ export class UserService {
   }
 
   async deleteUser(id: number) {
+    const user = await this.userRepository.findOneBy({ id });
+    if (!user) throw new NotFoundException('User not found');
+
     await this.userRepository.delete(id);
     await this.refreshTokenRepository.delete({ userId: id });
